Abs int float

+++ ESP_Hover_Browser_GY_521/ESP_Hover_Browser_GY_521.ino
@@ -212,7 +212,7 @@ void updateMotors()
     */
      
       int z_motorsnelheid = map(ui_slider2,0,360,0,PWM_RANGE); // voor zweefmotor
-      if (abs((float)ui_joystick_y * (float)ui_joystick_x) < 5) {
+      if (abs((ui_joystick_y * ui_joystick_x) < 5) {
         z_motorsnelheid = 0; // bij joystick los ook zweefmotor uit
       }
